requirements.txt
numpy==1.22.3
pandas==1.4.2
pytorch_lightning==1.6.3
setuptools==61.2.0
tokenizers==0.12.1
tqdm==4.64.0
transformers==4.19.2
nervaluate==0.1.8
